refactor:  리코일 변경사항에 따른 generator 변경
